requirements/base.txt
fastapi==0.87.0
uvicorn==0.19.0
inflection==0.5.1
asyncpg==0.27.0
tortoise-orm==0.19.2
fastapi-utils==0.2.1
psycopg2-binary==2.9.5
python-dotenv==0.21.0
aerich==0.7.1
